requirements.txt
flask
gunicorn
python-binance
click
dnspython
Flask-PyMongo
itsdangerous
pymongo
apscheduler
flask_socketio
eventlet
asyncio
